Move constants and enums from housing to projectCommon for cross initiative use
