+++ frontend/src/components/electrification/project/ProjectIntakeForm.vue
@@ -8,8 +8,7 @@ import { useRouter } from 'vue-router';
 import AdvancedFileUpload from '@/components/file/AdvancedFileUpload.vue';
 import Divider from '@/components/common/Divider.vue';
 import { AutoComplete, FormAutosave, FormNavigationGuard, InputText, RadioList, TextArea } from '@/components/form';
-import ContactCardIntakeForm from '@/components/form/common/ContactCardIntakeForm.vue';
-import CollectionDisclaimer from '@/components/housing/CollectionDisclaimer.vue';
+import { CollectionDisclaimer, ContactCardIntakeForm } from '@/components/form/common';
 import { createProjectIntakeSchema } from '@/components/electrification/project/ProjectIntakeSchema';
 import { Button, Card, Message, useConfirm, useToast } from '@/lib/primevue';
 import { documentService, electrificationProjectService, externalApiService } from '@/services';
