+++ frontend/src/components/form/common/index.ts
@@ -1,0 +1,5 @@
+export { default as CollectionDisclaimer } from './CollectionDisclaimer.vue';
+export { default as ContactCardIntakeForm } from './ContactCardIntakeForm.vue';
+export { default as ContactCardNavForm } from './ContactCardNavForm.vue';
+export { default as LocationCard } from './LocationCard.vue';
+export { default as NaturalDisasterCard } from './NaturalDisasterCard.vue';
