+++ frontend/src/utils/constants/housing.ts
@@ -1,13 +1,7 @@
 /*
  * Housing initiative constants
  */
-import {
-  ContactPreference,
-  NumResidentialUnits,
-  ProjectApplicant,
-  ProjectLocation,
-  ProjectRelationship
-} from '../enums/housing';
+import { NumResidentialUnits, ProjectApplicant, ProjectLocation } from '../enums/housing';
 
 export const ADDRESS_CODER_QUERY_PARAMS = {
   echo: false,
@@ -20,12 +14,6 @@ export const ADDRESS_CODER_QUERY_PARAMS = {
   matchPrecision: 'occupant, unit, site, civic_number, intersection, block, street, locality, province',
   precisionPoints: 100
 };
-
-export const CONTACT_PREFERENCE_LIST = [
-  ContactPreference.PHONE_CALL,
-  ContactPreference.EMAIL,
-  ContactPreference.EITHER
-];
 
 export const HOUSING_ASSISTANCE = {
   email: 'Navigator.Service@gov.bc.ca',
@@ -49,12 +37,6 @@ export const ORG_BOOK_QUERY_PARAMS = {
 };
 
 export const PROJECT_APPLICANT_LIST = [ProjectApplicant.BUSINESS, ProjectApplicant.INDIVIDUAL];
-
-export const PROJECT_RELATIONSHIP_LIST = [
-  ProjectRelationship.OWNER,
-  ProjectRelationship.CONSULTANT,
-  ProjectRelationship.OTHER
-];
 
 export const PROJECT_LOCATION_LIST = [
   ProjectLocation.LOCATION_COORDINATES,
